requirements.txt
google-genai
google-cloud-storage